+++ StreamDeckVSC/Keys/ExecuteTerminalCommandKey.cs
@@ -7,7 +7,7 @@ namespace StreamDeckVSC.Keys
     [PluginActionId("com.nicollasr.streamdeckvsc.executeterminalcommand")]
     public class ExecuteTerminalCommandKey : KeyBase<TerminalCommandSettings>
     {
-        public ExecuteTerminalCommandKey(SDConnection connection, InitialPayload payload) : base(connection, payload)
+        public ExecuteTerminalCommandKey(ISDConnection connection, InitialPayload payload) : base(connection, payload)
         {
         }